move get_llm_models function
